requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
azure-core
azure-ai-documentintelligence
python-multipart